New dials.resolutionizer dispatcher
- Keep xia2.resolutionizer synonym for now
- Replace print/stamp calls with logging calls
